+++ ReactiveUIToolKitDocs~/src/pages/Components/RadioButton/RadioButtonPage.tsx
@@ -19,13 +19,13 @@ export const RadioButtonPage: FC = () => (
       <Typography variant="h5" component="h2" gutterBottom>
         Props
       </Typography>
-      <CodeBlock language="tsx" code={getPropsDoc('RadioButtonProps')} />
+      <CodeBlock language="jsx" code={getPropsDoc('RadioButtonProps')} />
     </Box>
     <Box sx={Styles.section}>
       <Typography variant="h5" component="h2" gutterBottom>
         Basic usage
       </Typography>
-      <CodeBlock language="tsx" code={RADIO_BUTTON_BASIC} />
+      <CodeBlock language="jsx" code={RADIO_BUTTON_BASIC} />
     </Box>
     <UnityDocsSection componentName="RadioButton" />
   </Box>