+++ ReactiveUIToolKitDocs~/src/pages/Components/Tab/TabPage.tsx
@@ -20,13 +20,13 @@ export const TabPage: FC = () => (
       <Typography variant="h5" component="h2" gutterBottom>
         Props
       </Typography>
-      <CodeBlock language="tsx" code={getPropsDoc('TabProps')} />
+      <CodeBlock language="jsx" code={getPropsDoc('TabProps')} />
     </Box>
     <Box sx={Styles.section}>
       <Typography variant="h5" component="h2" gutterBottom>
         Basic usage
       </Typography>
-      <CodeBlock language="tsx" code={TAB_BASIC} />
+      <CodeBlock language="jsx" code={TAB_BASIC} />
     </Box>
     <UnityDocsSection componentName="Tab" />
   </Box>